refac: isolate components
