+++ src/App.tsx
@@ -1,10 +1,14 @@
 import React from "react";
-import ProfileHeader from "./components/profileHeader/ProfileHeader";
+import Main from "./pages/Main";
 
 const App = () => {
   return (
     <div className="App">
-      <ProfileHeader />
+      <section className="bg-white dark:bg-gray-900">
+        <div className="flex flex-col justify-between items-center gap-4 p-5 md:mx-56 lg:py-16 lg:px-6 ">
+          <Main />
+        </div>
+      </section>
     </div>
   );
 };
